Add files via upload
For Task 2 of the internship, I successfully completed the development of a stopwatch web application. This task required creating a functional and visually appealing stopwatch that allows users to start, stop, and reset the timer, with the ability to record lap times. The project involved using HTML, CSS, and JavaScript to ensure accurate time tracking and a seamless user experience. Special attention was given to the application's responsiveness and performance, making it efficient and accessible across different devices.